[Platinum V] Title: 팩토리얼과 점화식, Time: 1616 ms, Memory: 406044 KB -BaekjoonHub

+++ 백준/Platinum/13439. 팩토리얼과 점화식/팩토리얼과 점화식.cc
@@ -2,14 +2,14 @@
 using namespace std;
 typedef long long ll;
 const ll mod=1e9+9;
-unordered_map<ll,ll> dp[101][1001];
-unordered_map<ll,ll> go(ll n,ll k){
+unordered_map<int,int> dp[101][1001];
+unordered_map<int,int> go(int n,int k){
     if(!n) return {};
     if(!dp[k][n].empty())return dp[k][n];
     if(!k){
-        unordered_map<ll,ll> ma;
-        ll nn=n;
-        for(ll i=2;i*i<=nn;i++)while(n%i==0)ma[i]++,n/=i;
+        unordered_map<int,int> ma;
+        int nn=n;
+        for(int i=2;i*i<=nn;i++)while(n%i==0)ma[i]++,n/=i;
         if(n>1)ma[n]++;
         return dp[k][nn]=ma;
     }
@@ -23,6 +23,6 @@ int main(){
     ll n,k;scanf("%lld%lld",&n,&k);
     ll ans=1;
     for(auto [a,b]:go(n,k))
-        ans=(ans*(b+1)%mod)%mod;
+        ans=(ans*(ll)(b+1)%mod)%mod;
     printf("%lld",ans);
 }
